Use explicit separate caches
This breaks up the one unified "api" cache into multiple explicit
caches for each build stage. I'm not sure how Travis handles caching
and if it can deal with only updating a portion of the cache.

This also changes the build directory name to be "build". This is more
explicit as it is the directory used for building APIs, but it doesn't
just contain APIs.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>