Build FIRRTL specification for each API version
Modify the build flow to, as part of building FIRRTL APIs, also build
the FIRRTL specification. The resulting specification is placed inside
the API directory for a specific tag. Entries on the FIRRTL section of
the website are then added (one for each specification version).

To build the FIRRTL spec, a script is
used (scripts/build-firrtl-spec.sh) which will build the FIRRTL
specification using the author date of the commit in which it is
included. This allows for the spec to be built in CI, but to get the
correct date.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>